Move `enum_extend_if_invalid` to `utils.py`
