+++ fbchat/utils.py
@@ -11,6 +11,7 @@ from os.path import basename
 import warnings
 import logging
 import requests
+import aenum
 from .models import *
 
 try:
@@ -297,3 +298,12 @@ def get_files_from_paths(filenames):
     yield files
     for fn, fp, ft in files:
         fp.close()
+
+
+def enum_extend_if_invalid(enumeration, value):
+    try:
+        return enumeration(value)
+    except ValueError:
+        log.warning("Failed parsing {}({!r}). Extending enum.".format(enumeration, value))
+        aenum.extend_enum(enumeration, "UNKNOWN_{}".format(value).upper(), value)
+        return enumeration(value)
